+++ include/mp++/real.hpp
@@ -33,7 +33,6 @@
 #include <vector>
 
 #include <mp++/concepts.hpp>
-#include <mp++/demangle.hpp>
 #include <mp++/detail/fwd_decl.hpp>
 #include <mp++/detail/gmp.hpp>
 #include <mp++/detail/mpfr.hpp>
@@ -45,6 +44,7 @@
 #if defined(MPPP_WITH_QUADMATH)
 #include <mp++/real128.hpp>
 #endif
+#include <mp++/type_name.hpp>
 
 namespace mppp
 {
@@ -1738,7 +1738,7 @@ private:
     template <typename T>
     [[noreturn]] void raise_overflow_error() const
     {
-        throw std::overflow_error("Conversion of the real " + to_string() + " to the type '" + demangle<T>()
+        throw std::overflow_error("Conversion of the real " + to_string() + " to the type '" + type_name<T>()
                                   + "' results in overflow");
     }
     // Unsigned integrals, excluding bool.